everything fails. no return.

+++ FYS4150_Project2/Math/factorial.cpp
@@ -1,9 +1,10 @@
 #include "factorial.h"
 #include <cassert>
+#include <stdio.h>
 
-int factorial(int n) {
+double factorial(int n) {
     assert(n>=0);
-    int fact = 1;
+    double fact = 1.;
     for (int i = 2; i <= n; i++) {
         fact *= i;
     }
